Use binary format in WasmExampleTests

+++ backend-wasm/src/main/kotlin/org/shedlang/compiler/backends/wasm/wasm/Leb128Encoding.kt
@@ -25,7 +25,7 @@ internal object Leb128Encoding {
         while (hasMore) {
             hasMore = (remaining != end
                 || remaining and 1 != value shr 6 and 1)
-            bytes.add((value and 0x7f or if (hasMore) 0x80 else 0).toUByte())
+            bytes.add((value and 0x7f or if (hasMore) 0x80 else 0).toByte().toUByte())
             value = remaining
             remaining = remaining shr 7
         }
